requirements_test.txt
pytest>=7.4.0
pytest-homeassistant-custom-component>=0.13.95
